Renamed to variable_transformation after some functionality removed
